+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - [x] Handle scenario where recalculate is called twice when total is updated *NOT REPRODUCIBLE*
 - [x] Ensure wording between delete/remove is consistent UX
 - [x] Add https://reactjs.org/docs/error-boundaries.html
+- [x] Protect names from bad string input